requirements.txt
anthropic
supabase
python-dotenv
slack_bolt
gunicorn